+++ Magic/CHANGELOG.md
@@ -27,6 +27,7 @@
      - Random: Randomly choose one populators from a distance-weighted probability map
      - Repeat: Repeat another populator some random amount of times
  - Use generators in world configs for completely custom world gen
+ - Editor support exists, use `/meditor populator` or `/meditor generator`
  - Allow modifying game rules of a world via `game_rules` config
  - Add `time` parameter to set a world to a specific time of day
  - Add `title_delay` parameter to show players a title when they enter a world